refactor of the consultation record & add consultation record proponent and staff
