+++ submit-web/src/components/Shared/YesNoRadioOptions.tsx
@@ -6,19 +6,35 @@ export const NO = false;
 
 type IYesNoRadioOptionsProps = {
   error: boolean;
+  disabled: boolean;
 };
 export const YesNoRadioOptions = ({
   error = true,
+  disabled = false,
 }: IYesNoRadioOptionsProps) => {
   const sx = [
+    disabled && {
+      color: BCDesignTokens.iconsColorDisabled,
+    },
     error && {
       color: BCDesignTokens.surfaceColorPrimaryDangerButtonDefault,
     },
   ];
+
   return (
     <>
-      <FormControlLabel value={YES} control={<Radio sx={sx} />} label="Yes" />
-      <FormControlLabel value={NO} control={<Radio sx={sx} />} label="No" />
+      <FormControlLabel
+        value={YES}
+        disabled={disabled}
+        control={<Radio sx={sx} disabled={disabled} />}
+        label="Yes"
+      />
+      <FormControlLabel
+        value={NO}
+        disabled={disabled}
+        control={<Radio sx={sx} disabled={disabled} />}
+        label="No"
+      />
     </>
   );
 };
