+++ submit-web/src/components/Shared/controlled/ControlledRadioGroup.tsx
@@ -9,11 +9,13 @@ import { Controller, useFormContext } from "react-hook-form";
 
 type IFormInputProps = {
   name: string;
+  disabled?: boolean;
 } & RadioGroupProps;
 
 const ControlledRadioGroup: FC<IFormInputProps> = ({
   name,
   children,
+  disabled,
   ...otherProps
 }) => {
   const {
@@ -28,8 +30,9 @@ const ControlledRadioGroup: FC<IFormInputProps> = ({
         control={control}
         name={name}
         defaultValue={defaultValues?.[name] || ""}
+        disabled={disabled}
         render={({ field }) => (
-          <RadioGroup {...otherProps} {...field}>
+          <RadioGroup disabled={disabled} {...otherProps} {...field}>
             {children}
           </RadioGroup>
         )}
