+++ submit-web/src/components/SubmissionItem/ConsultationRecord/ConsultationRecordProponentView/constants.ts
@@ -1,0 +1,3 @@
+export const CONSULTATION_RECORD_DOCUMENT_FOLDERS = Object.freeze({
+  CONSULTATION_RECORDS: "consultation_records",
+});
